src/memory/requirements.txt
# Memory Module Dependencies

# Vector Database
patann>=0.1.3
sentence-transformers>=2.0.0

# Graph Database
neo4j>=5.0

# Utilities
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
